[refactor] component구조 변경
- ui를 담당하는 폴더와 여러 ui에 중복적으로 사용할 수 있는 layout을 담당하는 폴더로 분리
