+++ src/pages/main.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import Map from '../components/Map';
-import Marker from '../components/Marker';
-import Header from '../components/Header';
+import Map from '../components/UI/Map';
+import Marker from '../components/UI/Marker';
+import Header from '../components/UI/Header';
 
 // TODO: remove
 const dummy = [{ lng: -86.89871737888747, lat: 40.41866254968954 }];
